add webservice_tarball.sh
Since Quay.io cannot build images if the Dockerfile exists in a subtree
but the rest of the build context does not, it's necessary to take this
additional step of copying the Dockerfile into the root. This script
does that and generates a tarball which can then be uploaded.

Also, removing Dockerfile from .dockerignore is necessary.